[Docs] First iteration of documentation for `Gateway`s and the  `Claim & Proof Lifecycle` (#393)
## Summary

This PR aims to set a foundation for a rough initial draft of the documentation as it relates to `Gateways` and the `Claim & Proof Lifecycle`.

## Context

This PR DOES NOT aim to be an exhaustive, easy to follow and reviewable final version of the documentation due to all the priorities happening as of submitting it. 

My request to the reviewer (@red-0ne) is to see the TODOs and consider a "blind approval" so it can be merged and deployed to dev.poktroll.com until there is time to review it in depth at a future point in time. For example:

```md
TODO(@Olshansk): Iterate on this doc & link to governance params.

TODO(@red-0ne): Review this document and submit a PR with updates & edits.
```

As additional context, the [upcoming presentation](https://docs.google.com/presentation/d/1xlCGzS_oHXJOzvcu-jHZUfmhD3qeVCzc6SUSJijTuJ4/edit#slide=id.g2ba566bc8cf_2_0) will serve as a good foundation of how to update the documentation,