[Bug] Nil txResponse on Supplier client (#1402)
## Summary

This pull request adds null checks for `txResponse` in logging
statements within the `SubmitProofs` and `CreateClaims` methods of
`pkg/client/supplier/client.go`.

These changes ensure that logging operations are only performed when
`txResponse` is not nil, improving the robustness of the code.

## Issue


![image](https://github.com/user-attachments/assets/bce6d204-4f6c-4aa0-957f-4887c879867e)

## Type of change

Select one or more from the following:

- [ ] New feature, functionality or library
- [x] Bug fix
- [ ] Code health or cleanup
- [ ] Documentation
- [ ] Other (specify)

## Sanity Checklist

- [x] I have updated the GitHub Issue Metadata: `assignees`,
`reviewers`, `labels`, `project`, `iteration` and `milestone`
- [ ] For docs: `make docusaurus_start`
- [x] For small changes: `make go_develop_and_test` and `make test_e2e`
- [ ] For major changes: `devnet-test-e2e` label to run E2E tests in CI
- [ ] For migration changes: `make test_e2e_oneshot`
- [ ] 'TODO's, configurations and other docs