[Documentation] First attempt at adding chat to docs (#1472)
Experiment with an LLM chatbot that embeds directly into docusuarus.

The chat itself is quite shitty.

The goal of this PR was to get the end-to-end deployment in place.

The [original repo](https://github.com/nichnarmada/docusaurus-plugin-chat-page) was [forked](https://github.com/buildwithgrove/docusaurus-plugin-chat-page/commit/158fc769eed2f11b71b8630f8552d22a0e06a4a3) so we can iterate on it.

The first feature add in the new repo was "cached embeddings".